Move in memory integration tests into the main Api.Tests project
