+++ tests/Api.Tests/Endpoints/CustomsDeclarations/GetTests.cs
@@ -1,7 +1,9 @@
 using System.Net;
 using Defra.TradeImportsDataApi.Api.Services;
+using Defra.TradeImportsDataApi.Api.Tests;
+using Defra.TradeImportsDataApi.Api.Tests.Endpoints;
 using Defra.TradeImportsDataApi.Data.Entities;
-using Defra.TradeImportsDataApi.Domain.Ipaffs;
+using Defra.TradeImportsDataApi.Domain.CustomsDeclaration;
 using Defra.TradeImportsDataApi.Testing;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
@@ -9,14 +11,14 @@ using NSubstitute;
 using WireMock.Server;
 using Xunit.Abstractions;
 
-namespace Defra.TradeImportsDataApi.Api.IntegrationTests.Endpoints.ImportPreNotifications;
+namespace Defra.TradeImportsDataApi.Api.IntegrationTests.Endpoints.CustomsDeclarations;
 
 public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
 {
-    private IImportPreNotificationService MockImportPreNotificationService { get; } =
-        Substitute.For<IImportPreNotificationService>();
+    private ICustomsDeclarationService MockCustomsDeclarationService { get; } =
+        Substitute.For<ICustomsDeclarationService>();
     private WireMockServer WireMock { get; }
-    private const string ChedId = "chedId";
+    private const string Mrn = "mrn";
     private readonly VerifySettings _settings;
 
     public GetTests(ApiWebApplicationFactory factory, ITestOutputHelper outputHelper, WireMockContext context)
@@ -34,7 +36,7 @@ public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
     {
         base.ConfigureTestServices(services);
 
-        services.AddTransient<IImportPreNotificationService>(_ => MockImportPreNotificationService);
+        services.AddTransient<ICustomsDeclarationService>(_ => MockCustomsDeclarationService);
     }
 
     [Fact]
@@ -42,7 +44,7 @@ public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
     {
         var client = CreateClient();
 
-        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.ImportPreNotifications.Get(ChedId));
+        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.CustomsDeclarations.Get(Mrn));
 
         await VerifyJson(await response.Content.ReadAsStringAsync(), _settings);
     }
@@ -51,21 +53,20 @@ public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
     public async Task Get_WhenFound_ShouldReturnContent()
     {
         var client = CreateClient();
-        MockImportPreNotificationService
-            .GetImportPreNotification(ChedId, Arg.Any<CancellationToken>())
+        MockCustomsDeclarationService
+            .GetCustomsDeclaration(Mrn, Arg.Any<CancellationToken>())
             .Returns(
-                new ImportPreNotificationEntity
+                new CustomsDeclarationEntity
                 {
-                    Id = ChedId,
-                    CustomsDeclarationIdentifier = ChedId,
-                    ImportPreNotification = new ImportPreNotification(),
+                    Id = Mrn,
+                    ClearanceRequest = new ClearanceRequest(),
                     Created = new DateTime(2025, 4, 3, 10, 0, 0, DateTimeKind.Utc),
                     Updated = new DateTime(2025, 4, 3, 10, 15, 0, DateTimeKind.Utc),
                     ETag = "etag",
                 }
             );
 
-        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.ImportPreNotifications.Get(ChedId));
+        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.CustomsDeclarations.Get(Mrn));
 
         await VerifyJson(await response.Content.ReadAsStringAsync(), _settings)
             .UseMethodName(nameof(Get_WhenFound_ShouldReturnContent));
@@ -77,7 +78,7 @@ public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
     {
         var client = CreateClient(addDefaultAuthorizationHeader: false);
 
-        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.ImportPreNotifications.Get(ChedId));
+        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.CustomsDeclarations.Get(Mrn));
 
         response.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
     }
@@ -87,7 +88,7 @@ public class GetTests : EndpointTestBase, IClassFixture<WireMockContext>
     {
         var client = CreateClient(testUser: TestUser.WriteOnly);
 
-        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.ImportPreNotifications.Get(ChedId));
+        var response = await client.GetAsync(TradeImportsDataApi.Testing.Endpoints.CustomsDeclarations.Get(Mrn));
 
         response.StatusCode.Should().Be(HttpStatusCode.Forbidden);
     }
