+++ tests/Api.Tests/Endpoints/Gmrs/GetTests.Get_WhenBadRequest_ShouldBeBadRequest.verified.txt
