+++ tests/Api.Tests/Endpoints/Gmrs/GetTests.cs
@@ -1,5 +1,7 @@
 using System.Net;
 using Defra.TradeImportsDataApi.Api.Services;
+using Defra.TradeImportsDataApi.Api.Tests;
+using Defra.TradeImportsDataApi.Api.Tests.Endpoints;
 using Defra.TradeImportsDataApi.Data.Entities;
 using Defra.TradeImportsDataApi.Domain.Gvms;
 using Defra.TradeImportsDataApi.Testing;
