+++ tests/Api.Tests/Endpoints/Health/GetTests.Get_Health_ReturnsAnonymous.verified.txt
